Testing out score()
